+++ src/main/java/command/ExitCommand.java
@@ -1,3 +1,10 @@
+package command;
+
+import exception.UserInputException;
+import storage.Storage;
+import tasklist.TaskList;
+import ui.UI;
+
 public class ExitCommand extends Command {
     @Override
     public void execute(TaskList tasks, Storage fm) throws UserInputException {
