+++ src/main/java/task/Task.java
@@ -1,5 +1,7 @@
+package task;
+
 public abstract class Task {
-    protected String description;
+    public String description;
     protected boolean isDone;
 
     public Task(String description) {
